+++ include/leo/leo_model.h
@@ -48,8 +48,8 @@ public:
        const uint8_t atmo_gases_attenuation,
        const uint8_t precipitation_attenuation,
        const uint8_t enable_link_margin,
-       const float surface_watervap_density = 7.5,
-       const float temperature = 0, const float rainfall_rate = 25);
+       const double surface_watervap_density = 7.5,
+       const double temperature = 0, const double rainfall_rate = 25);
 
   leo_model()
   {
