+++ include/leo/parabolic_reflector_antenna.h
@@ -52,9 +52,9 @@ public:
    * \return a boost::shared_ptr to the constructed tracker object.
    */
   static generic_antenna::generic_antenna_sptr
-  make(uint8_t type, float frequency, int polarization, float pointing_error,
-       float diameter,
-       float aperture_efficiency);
+  make(uint8_t type, double frequency, int polarization, double pointing_error,
+       double diameter,
+       double aperture_efficiency);
 
   parabolic_reflector_antenna() {};
