+++ include/leo/quadrifilar_helix_antenna.h
@@ -50,8 +50,8 @@ public:
    * \return a boost::shared_ptr to the constructed tracker object.
    */
   static generic_antenna::generic_antenna_sptr
-  make(uint8_t type, float frequency, int polarization,
-       float pointing_error, float loop);
+  make(uint8_t type, double frequency, int polarization,
+       double pointing_error, double loop);
 
   quadrifilar_helix_antenna() {};
