+++ include/leo/white_gaussian_noise.h
@@ -42,7 +42,8 @@ public:
   ~white_gaussian_noise();
 
   void
-  add_noise(gr_complex *outbuf, const gr_complex *inbuf, size_t num, float power);
+  add_noise(gr_complex *outbuf, const gr_complex *inbuf, size_t num,
+            double power);
 
 private:
   gr::random d_rng;
